fragments - first version

+++ app/src/main/java/com/example/bublesortgame/Model/Fragment.kt
@@ -3,7 +3,7 @@ package com.example.bublesortgame.Model
 import kotlin.random.Random
 
 
-class Fragment(var X: Float, var Y: Float, val tX: Float, val tY: Float, val duration: Int, val colour: Colour, val diam: Float, val id: Int = Random.nextInt())
+class Fragment(var X: Float, var Y: Float, val tX: Float, val tY: Float, val duration: Long, val colour: Colour, val diam: Float, val id: Int = Random.nextInt())
 {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
